Check-box fields fixed. Rooms controller was created and associated with Room operations. The _form file was created and the DRY method was followed.
