--- app/views/admin/add_room.html.erb
@@ -1,63 +1,0 @@
-<div class="container">
-  <div id="content-wrapper" class="d-flex flex-column">
-
-            <!-- Main Content -->
-            <div id="content">
-
-                <div class="container-fluid">
-
-                    <!-- Page Heading -->
-                    <h1 class="h3 mb-2 text-gray-800">Yeni Oda Ekle</h1>
-                    
-
-                    <div class="card shadow mb-4">
-                        <div class="card-header py-3">
-                          <div class="row container">
-                            <div class="col-md-6" style="margin: auto;">
-                                <h6 class="m-0 font-weight-bold text-primary">Oda Bilgileri</h6>
-                                  <%= render 'layouts/errors' %>
-                            </div>
-                            <div class="col-md-6">
-                                <%= button_to "Geri", room_path, method: :get, class: "btn btn-danger float-right" %>        
-                            </div>  
-                          </div>
-                        </div>
-                        
-                        <div class="card-body">
-                            <div class="table-responsive">
-                              <%= form_for @room, url: {action: :add_room}, html: {'data-turbo' => "false"} do |form| %>
-                                <div class="container">
-                                  <%= form.label "Oda Numarası", class: "form-check-label" %>
-                                  <%= form.number_field :number, placeholder: '42', class: "form-control", pattern: "[0-9.]+" %> <br>
-                                  <%= form.label "Oda Kapasitesi", class: "form-check-label" %>
-                                  <%= form.number_field :capacity, placeholder: '2', class: "form-control" %> <br>
-                                  <%= form.label "Oda Fiyatı", class: "form-check-label" %>
-                                  <%= form.text_field :price, placeholder: '599', class: "form-control" %> <br>
-                                  <%= form.label "Oda Açıklaması", class: "form-check-label" %>
-                                  <%= form.text_field :description, placeholder: 'Bu oda 2 kişiliktir. Televizyon ....', class: "form-control" %> <br>
-                                  <%= form.label "Oda Durumu", class: "form-check-label" %> <br>
-                                  <%= form.select :status, options_for_select([['Oda Boş', 'Müsait'], ['Oda Dolu', 'Bakımda']]) %>
-                                  <br><br>
-                                  <div>
-                                    <%= hidden_field_tag "room[service_ids][]", nil %> 
-                                    <% Service.all.each do |service| %>
-                                      <%= check_box_tag "room[service_ids][]", service.id, @room.service_ids.include?(service.id), id: dom_id(service) %>
-                                      <%= label_tag dom_id(service), service.name %>
-                                    <% end %>
-                                  </div>
-                                  <br>
-                                  <%= form.label "Fotoğraflar", class: "form-check-label" %> <br>
-                                  <div class="custom-file">
-                                    <%= form.file_field :photos, multiple: true %> <br>
-                                  </div>
-                                  <br> <br>
-                                  <%= form.submit 'Oda Oluştur', class: "btn btn-primary" %>
-                                </div>
-                              <% end %>
-                            </div>
-                        </div>
-                    </div>
-
-    
-  </div>
-</div>
