--- app/views/admin/room.html.erb
@@ -1,81 +1,0 @@
-<div class="container">
-    <div id="content-wrapper" class="d-flex flex-column">
-
-        <!-- Main Content -->
-        <div id="content">
-
-            <div class="container-fluid">
-                <!-- Page Heading -->
-                <h1 class="h3 mb-2 text-gray-800">Rooms</h1>
-
-                    <!-- DataTales Example -->
-                    <div class="card shadow mb-4">
-                        <div class="card-header py-3">
-                          <div class="row container">
-                            <div class="col-md-6" style="margin: auto;">
-                                <h6 class="m-0 font-weight-bold text-primary">All Of Rooms</h6>
-                            </div>
-                            <div class="col-md-6">
-                                <%= button_to "Ekle", room_new_path, method: :get, class: "btn btn-primary float-right" %>                                
-                            </div>  
-                          </div>
-                        </div>
-                        
-                        <div class="card-body">
-                            <div class="table-responsive">
-                                <table class="table table-bordered" id="dataTable" width="100%" cellspacing="0">
-                                    <thead>
-                                        <tr>
-                                            <th>Number</th>
-                                            <th>Capacity</th>
-                                            <th>Price</th>
-                                            <th>Description</th>
-                                            <th>Photos</th>
-                                            <th>Status</th>
-                                            <th>Services</th>
-                                            <th colspan="2">Actions</th>
-                                        </tr>
-                                    </thead>
-
-                                    <tbody>
-                                    
-                                        <% @rooms.each do |room| %>
-                                            <tr>
-                                                <td><%= room.number %></td>
-                                                <td><%= room.capacity %></td>
-                                                <td><%= room.price %></td>
-                                                <td><%= short_descripton(room.description)%></td>
-                                                <td>                                           
-                                                    <% room.photos.each do |photo| %>
-                                                        <%= link_to image_tag(room.photo_as_thumbnail(photo)), photo %>
-                                                    <% end %>
-                                                </td>
-                                                <td><%= room.status %></td>
-                                                <td><% room.services.pluck(:icon).each do |icon| %>
-                                                    <%= icon.html_safe %>  
-                                                    <% end %>
-                                                </td>
-                                                <td>
-                                                    <%= button_to "Düzenle", edit_room_path(room.id), method: :get, class: "btn btn-info float-right" %> 
-                                                </td>
-                                                <td>
-                                                    <%= button_to "Sil", delete_room_path(room.id), method: :delete, class: "btn btn-danger float-right" %>
-                                                </td>
-                                            </tr>
-                                        <% end %>
-                                            
-                                        
-                                    </tbody>
-                                </table>
-                            </div>
-                        </div>
-                    </div>
-
-            </div>
-            <!-- /.container-fluid -->
-
-        </div>
-        <!-- End of Main Content -->
-
-    </div>
-</div>
