+++ app/views/admin/shared/_sidebar.html.erb
@@ -29,7 +29,7 @@
 
             <!-- Nav Item - Charts -->
             <li class="nav-item">
-                <%= link_to(room_path, class: "nav-link") do %>
+                <%= link_to(rooms_path, class: "nav-link") do %>
                     <i class="fas fa-fw fa-chart-area"></i>
                     <span>Oda İşlemleri</span></a>
                 <% end %>
