+++ app/views/rooms/edit.html.erb
@@ -1,0 +1,35 @@
+<div class="container">
+  <div id="content-wrapper" class="d-flex flex-column">
+
+            <!-- Main Content -->
+            <div id="content">
+
+                <div class="container-fluid">
+
+                    <!-- Page Heading -->
+                    <h1 class="h3 mb-2 text-gray-800">Oda Düzenle</h1>
+                    
+
+                    <div class="card shadow mb-4">
+                        <div class="card-header py-3">
+                          <div class="row container">
+                            <div class="col-md-6" style="margin: auto;">
+                                <h6 class="m-0 font-weight-bold text-primary">Oda Bilgileri</h6>
+                                <%= render 'layouts/errors' %>   
+                            </div>
+                            <div class="col-md-6">
+                                <%= button_to "Geri", rooms_path, method: :get, class: "btn btn-danger float-right" %>        
+                            </div>  
+                          </div>
+                        </div>
+                        
+                        <div class="card-body">
+                            <div class="table-responsive">
+                              <%= render 'rooms/shared/form' %>
+                            </div>
+                        </div>
+                    </div>
+
+    
+  </div>
+</div>
