+++ config/routes.rb
@@ -7,15 +7,7 @@ Rails.application.routes.draw do
   #rooms routes
   root "admin#index"
 
-  get 'room', to: 'admin#room'
-  get 'room/new', to: 'admin#add_room'
-  post 'room/new', to: 'admin#create_room'
-
-  get 'room/:id/edit', to: 'admin#room_edit', as: :edit_room
-  patch 'room/:id', to: 'admin#room_update'
-
-  delete 'room/:id', to: 'admin#room_destroy', as: :delete_room
-
+  resources :rooms
 
   #service routes
   get 'services', to: 'services#index'
